test(workflows): split the inverse tests to match the modules they cover
inverse_test.py spanned four of the five modules, and the evidence was in its own import block:
splitting the source had forced it to reach into .config and .report directly to keep running.
That was the file telling me it no longer matched anything.

    config_test.py   233   schema rejection: targets, scenarios, paths, seeds, accelerator
    seeds_test.py    101   selection policies, weighting, dedup, ranking
    report_test.py    73   the text formatting the seed→optimised map lays out by hand
    engine_test.py    77   end-to-end: a real run over the synthetic catalog
    conftest.py      202   the synthetic four-task catalog every one of them builds on

Verified the same way as the source split: all 41 definitions token-identical, decorators
included, none lost and none added — and the decorator check is there because slicing by lineno
dropped five @dataclass lines the last time it was skipped. 639 tests still pass, same as before.

This also completes what the earlier commits left inconsistent: models/inverse_design already had
per-module tests inside the package, so the convention existed and only workflows had not caught up.
